test: move firestore tests into __tests__ to match jest config
